+++ HISTORY.rst
@@ -7,6 +7,8 @@ PlatformIO 3.0
 3.1.0 (2016-09-??)
 ~~~~~~~~~~~~~~~~~~
 
+* New! Dynamic variables/templates for `Project Configuration File "platformio.ini" <http://docs.platformio.org/en/stable/projectconf.html>`__
+  (`issue #705 <https://github.com/platformio/platformio/issues/705>`_)
 * Implemented LocalCache system for API and improved a work in off-line mode
 * Improved Project Generator when custom ``--project-option`` is passed to
   `platformio init <http://docs.platformio.org/en/stable/userguide/cmd_init.html>`__